// src/game/Sim/Critters/types.d.ts
export declare enum CritterType {
    Beast = 0,
    Spirit = 1,
    Plant = 2,
    Fire = 3,
    Water = 4,
    Earth = 5,
    Air = 6,
    Metal = 7
}
export declare const TYPE_COUNT = 8;
export declare const TYPE_NAMES: Record<CritterType, string>;
export declare const TYPE_COLOR: Record<CritterType, number>;